# backend/internal/app/router.go
package app

import (
	"bufio"
	"fmt"
	"log"
	"net"
	"net/http"
	"strings"
	"time"

	"github.com/bms-del112/wuzz-chat/internal/api"
	"github.com/bms-del112/wuzz-chat/internal/auth"
	"github.com/bms-del112/wuzz-chat/internal/shared/ratelimit"
	"github.com/bms-del112/wuzz-chat/internal/storage"
)

// setupRouter mendaftarkan seluruh rute HTTP & WebSocket mux yang dikelompokkan
// rapi berdasarkan domain fungsional.
func (a *Application) setupRouter() http.Handler {
	mux := http.NewServeMux()

	// Helper CORS Middleware untuk REST API
	withCORS := func(h http.HandlerFunc) http.HandlerFunc {
		return func(w http.ResponseWriter, r *http.Request) {
			a.CorsValidator.Middleware(h).ServeHTTP(w, r)
		}
	}

	// =========================================================================
	// 1. HEALTH CHECK & PUBLIC CONFIGURATION
	// =========================================================================
	mux.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		fmt.Fprintf(w, `{"status":"ok","time":"%s"}`, time.Now().UTC().Format(time.RFC3339))
	})

	// Info versi aplikasi mobile (publik, tanpa auth; dikecualikan dari gatekeeper versi)
	if a.Config != nil {
		versionInfo := api.NewVersionInfoHandler(
			a.Config.MinMobileBuild, a.Config.LatestMobileBuild, a.Config.LatestMobileVersion,
			a.Config.MobileReleaseNotes, a.Config.PlayStoreURL, a.Config.AppStoreURL, a.Config.APKDownloadURL,
		)
		mux.HandleFunc(api.VersionInfoPath, withCORS(versionInfo.ServeHTTP))
	}

	if a.MediaHandler != nil {
		mux.HandleFunc("/api/config", withCORS(a.MediaHandler.Config))
	}

	// =========================================================================
	// 1B. OPENAPI SPECIFICATION & API DOCUMENTATION (Milestone 6)
	// =========================================================================
	if a.OpenAPIHandler != nil {
		mux.HandleFunc("/api/openapi.yaml", withCORS(a.OpenAPIHandler.ServeOpenAPISpec))
		mux.HandleFunc("/api/docs", withCORS(a.OpenAPIHandler.ServeDocsUI))
	}

	// =========================================================================
	// 2. STATIC UPLOADS & MEDIA ATTACHMENTS
	// =========================================================================
	uploadDir := a.Config.UploadDir
	if uploadDir == "" {
		uploadDir = "./uploads"
	}
	_ = storage.EnsureDir(uploadDir)
	fileServer := http.FileServer(api.NewUploadsFileSystem(uploadDir))
	fileHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Cache-Control", "public, max-age=86400")
		fileServer.ServeHTTP(w, r)
	})
	mux.Handle("/uploads/", http.StripPrefix("/uploads/", api.UploadsSecurityHeaders(a.CorsValidator.Middleware(fileHandler))))

	if a.MediaHandler != nil {
		// Kuota per user (batas longgar agar burst kirim banyak gambar/voice note tetap aman)
		uploadLimiter := ratelimit.NewIPRateLimiter(40, time.Minute)
		mux.HandleFunc("/api/media/upload", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(api.UserRateLimit(uploadLimiter)(http.HandlerFunc(a.MediaHandler.Upload))).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/media/signed-upload-url", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(api.UserRateLimit(uploadLimiter)(http.HandlerFunc(a.MediaHandler.CreateSignedUploadURL))).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/media/ack", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.MediaHandler.AcknowledgeDownload)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 3. LINK PREVIEW SCRAPER
	// =========================================================================
	if a.LinkPreviewHandler != nil {
		mux.HandleFunc("/api/link-preview", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.LinkPreviewHandler.ServeHTTP)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 4. AUTHENTICATION & IDENTITY (Dual-Tier Rate Limited)
	// =========================================================================
	if a.AuthHandler != nil {
		mux.HandleFunc("/api/auth/register", withCORS(func(w http.ResponseWriter, r *http.Request) {
			ratelimit.DualRateLimitMiddleware(a.AuthLimiter)(http.HandlerFunc(a.AuthHandler.Register)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/login", withCORS(func(w http.ResponseWriter, r *http.Request) {
			ratelimit.DualRateLimitMiddleware(a.AuthLimiter)(http.HandlerFunc(a.AuthHandler.Login)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/me", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.Me)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/refresh", withCORS(func(w http.ResponseWriter, r *http.Request) {
			// Tanpa AuthLimiter: bucket itu milik login/register (per IP) dan berbagi NAT akan memblokir login. Endpoint ini
			// murah (token segar → refreshed:false) dan wajib JWT sah.
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.Refresh)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/logout", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.Logout)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/sessions/revoke-others", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.RevokeAllOtherSessions)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/sessions", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.GetActiveSessions)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/sessions/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.RevokeSession)).ServeHTTP(w, r)
		}))
		if a.DeviceHandler != nil {
			mux.HandleFunc("/api/auth/devices", withCORS(func(w http.ResponseWriter, r *http.Request) {
				auth.RequireJWT()(http.HandlerFunc(a.DeviceHandler.ListDevices)).ServeHTTP(w, r)
			}))
			mux.HandleFunc("/api/auth/devices/", withCORS(func(w http.ResponseWriter, r *http.Request) {
				auth.RequireJWT()(http.HandlerFunc(a.DeviceHandler.RemoveDevice)).ServeHTTP(w, r)
			}))
		}
		if a.CredentialHandler != nil {
			mux.HandleFunc("/api/auth/credentials", withCORS(func(w http.ResponseWriter, r *http.Request) {
				auth.RequireJWT()(http.HandlerFunc(a.CredentialHandler.ListCredentials)).ServeHTTP(w, r)
			}))
		}
		mux.HandleFunc("/api/auth/profile", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.UpdateProfile)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/verify-password", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.VerifyPassword)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/change-password", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.ChangePassword)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/auth/public-key", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodGet || r.Method == http.MethodHead {
				if a.ChatHandler != nil {
					a.ChatHandler.GetUserPublicKey(w, r)
					return
				}
			}
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.UpdatePublicKey)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/user/public-key/reset", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.ResetPublicKey)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/users/public-key", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodGet || r.Method == http.MethodHead {
				if a.ChatHandler != nil {
					a.ChatHandler.GetUserPublicKey(w, r)
					return
				}
			}
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.UpdatePublicKey)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/users/public-key/reset", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.AuthHandler.ResetPublicKey)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 4B. EXTERNAL B2B & JIT PROVISIONING GATEWAY (Milestone 3)
	// =========================================================================
	if a.ProvisioningHandler != nil {
		mux.HandleFunc("/api/v1/auth/provision-token", withCORS(func(w http.ResponseWriter, r *http.Request) {
			api.B2BAuthGuard(a.TenantService)(http.HandlerFunc(a.ProvisioningHandler.ProvisionToken)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/v1/auth/exchange", withCORS(func(w http.ResponseWriter, r *http.Request) {
			a.ProvisioningHandler.ExchangeToken(w, r)
		}))
	}

	// =========================================================================
	// 5. E2EE DEVICE KEY TRANSFER (QR CODE)
	// =========================================================================
	if a.TransferHandler != nil {
		mux.HandleFunc("/api/users/transfer/create", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.TransferHandler.CreateSession)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/users/transfer/consume", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.TransferHandler.ConsumeSession)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 6. CHAT & MESSAGING (Direct, Pin, Forward, Receipt)
	// =========================================================================
	if a.ChatHandler != nil {
		mux.HandleFunc("/api/users/search", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.SearchUsers)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/users/profile", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.GetUserProfile)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/conversations", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.StartDirectChat)).ServeHTTP(w, r)
			} else if r.Method == http.MethodDelete {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.ClearConversation)).ServeHTTP(w, r)
			} else {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.GetConversations)).ServeHTTP(w, r)
			}
		}))
		mux.HandleFunc("/api/conversations/clear", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.ClearConversation)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/conversations/pin", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.PinConversation)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/conversations/unpin", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.UnpinConversation)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodGet {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.GetMessages)).ServeHTTP(w, r)
			} else if r.Method == http.MethodPut {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.EditMessage)).ServeHTTP(w, r)
			} else if r.Method == http.MethodDelete || r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.DeleteMessage)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/edit", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPut || r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.EditMessage)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/delete", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.DeleteMessage)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/messages/forward", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.ForwardMessage)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/receipt", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.UpdateReceipt)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/pin", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.PinMessage)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/unpin", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.UnpinMessage)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/pinned", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodGet {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.GetPinnedMessages)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/messages/search", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodGet {
				auth.RequireJWT()(http.HandlerFunc(a.ChatHandler.SearchMessages)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
	}

	// =========================================================================
	// 7. GROUP CHAT & FORUM ENGINE
	// =========================================================================
	if a.GroupHandler != nil {
		mux.HandleFunc("/api/groups", withCORS(func(w http.ResponseWriter, r *http.Request) {
			if r.Method == http.MethodPost {
				auth.RequireJWT()(http.HandlerFunc(a.GroupHandler.CreateGroup)).ServeHTTP(w, r)
			} else {
				http.Error(w, `{"error":"Method not allowed"}`, http.StatusMethodNotAllowed)
			}
		}))
		mux.HandleFunc("/api/groups/search", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.GroupHandler.SearchPublicGroups)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/groups/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.GroupHandler.RouteGroupRequest)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 8. GROUP MEMORY AI (DRAFT REVIEW & MEMBER KNOWLEDGE)
	// =========================================================================
	if a.MemoryHandler != nil {
		mux.HandleFunc("/api/memory/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.MemoryHandler.RouteMemoryRequest)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/memories/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.MemoryHandler.RouteApprovedMemoryRequest)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 9. PUSH NOTIFICATIONS (WEB PUSH VAPID)
	// =========================================================================
	if a.NotificationHandler != nil {
		mux.HandleFunc("/api/notifications/vapid-public-key", withCORS(a.NotificationHandler.GetVAPIDPublicKey))
		mux.HandleFunc("/api/notifications/subscribe", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.NotificationHandler.Subscribe)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/notifications/unsubscribe", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.NotificationHandler.Unsubscribe)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 9B. COMMUNITY SOCIAL FEED (Milestone M-Mobile-9.2)
	// =========================================================================
	if a.FeedHandler != nil {
		mux.HandleFunc("/api/feed", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.FeedHandler.HandleFeedRoot)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/feed/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.FeedHandler.RouteFeedRequest)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 9C. USER CONNECTIONS & FRIENDLIST ENGINE (Milestone M-Mobile-10)
	// =========================================================================
	if a.ConnectionHandler != nil {
		mux.HandleFunc("/api/connections", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ConnectionHandler.RouteConnectionRequest)).ServeHTTP(w, r)
		}))
		mux.HandleFunc("/api/connections/", withCORS(func(w http.ResponseWriter, r *http.Request) {
			auth.RequireJWT()(http.HandlerFunc(a.ConnectionHandler.RouteConnectionRequest)).ServeHTTP(w, r)
		}))
	}

	// =========================================================================
	// 10. WEBSOCKET REALTIME CONNECTION
	// =========================================================================
	if a.WsHandler != nil {
		mux.Handle("/ws", a.WsHandler)
	}

	var handler http.Handler = mux

	// Validasi versi aplikasi klien mobile (Force Update Gatekeeper)
	if a.Config != nil {
		versionMw := api.NewVersionMiddleware(a.Config.MinMobileBuild, a.Config.PlayStoreURL, a.Config.AppStoreURL).WithAPKURL(a.Config.APKDownloadURL)
		handler = versionMw.Middleware(handler)
	}

	// Selalu aktif: guard header/JWT tenant tidak boleh hilang hanya karena TenantService belum terpasang.
	// Tanpa TenantService, middleware tetap menyuntikkan tenant context tetapi melewati validasi registry.
	handler = api.NewTenantMiddleware(a.TenantService).Handler(handler)

	// Bungkus dengan request logger middleware di level terluar agar seluruh request tercatat
	handler = requestLoggerMiddleware(handler)

	return handler
}

// responseWriterRecorder membungkus http.ResponseWriter untuk mencatat HTTP status code dan bytes written.
// Mendukung http.Hijacker dan http.Flusher agar kompatibel dengan WebSocket upgrade (/ws) dan streaming.
type responseWriterRecorder struct {
	http.ResponseWriter
	statusCode   int
	bytesWritten int64
}

func (r *responseWriterRecorder) WriteHeader(statusCode int) {
	r.statusCode = statusCode
	r.ResponseWriter.WriteHeader(statusCode)
}

func (r *responseWriterRecorder) Write(b []byte) (int, error) {
	if r.statusCode == 0 {
		r.statusCode = http.StatusOK
	}
	n, err := r.ResponseWriter.Write(b)
	r.bytesWritten += int64(n)
	return n, err
}

func (r *responseWriterRecorder) Hijack() (net.Conn, *bufio.ReadWriter, error) {
	if hj, ok := r.ResponseWriter.(http.Hijacker); ok {
		return hj.Hijack()
	}
	return nil, nil, fmt.Errorf("http.Hijacker tidak didukung oleh underlying response writer")
}

func (r *responseWriterRecorder) Flush() {
	if fl, ok := r.ResponseWriter.(http.Flusher); ok {
		fl.Flush()
	}
}

// requestLoggerMiddleware mencatat seluruh lalu lintas HTTP yang masuk ke server backend.
func requestLoggerMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		rec := &responseWriterRecorder{
			ResponseWriter: w,
			statusCode:     http.StatusOK, // default jika handler tidak memanggil WriteHeader
		}

		next.ServeHTTP(rec, r)

		duration := time.Since(start)

		// Ambil client IP canonical
		ip := r.Header.Get("CF-Connecting-IP")
		if ip == "" {
			if xff := r.Header.Get("X-Forwarded-For"); xff != "" {
				if parts := strings.Split(xff, ","); len(parts) > 0 {
					ip = strings.TrimSpace(parts[0])
				}
			}
		}
		if ip == "" {
			ip = r.Header.Get("X-Real-IP")
		}
		if ip == "" {
			ip = r.RemoteAddr
		}

		ua := r.UserAgent()
		if len(ua) > 60 {
			ua = ua[:60] + "..."
		}

		log.Printf("[HTTP] %s %s %d %v | IP: %s | UA: %s", r.Method, r.URL.Path, rec.statusCode, duration.Round(time.Millisecond), ip, ua)
	})
}