# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WuzzChat Engine API
  version: 1.0.0
  description: |
    Enterprise Multi-Tenant Messaging Engine, Headless Realtime Chat, AI Memory, and E2EE Signaling Server.
    
    WuzzChat is a production-grade, headless communication engine designed for modern web, mobile (Flutter/React Native/Native iOS & Android), and B2B third-party platform integrations.
    
    Key Features:
    - **Multi-Tenancy**: Data isolation with automatic tenant resolution via `X-Tenant-ID` header or JWT claims.
    - **B2B Token Provisioning**: Server-to-server JIT provisioning via `X-App-ID` & `X-App-Secret` and 60-second client token exchange.
    - **Zero-Knowledge E2EE**: ECDH (P-256) + HKDF-SHA256 + AES-256-GCM architecture. Server never sees plaintext content.
    - **Level 2 Multi-Device Support**: Up to 2 concurrent devices per user with FIFO session eviction (`SESSION_REPLACED`).
    - **Store-and-Forward Media**: Ephemeral media auto-deletion upon client download acknowledgement (ACK) for 1-on-1 chats; 7-day retention for group fora.
    - **AI Memory Engine**: Human-in-the-Loop review lifecycle for AI-synthesized group memories and ephemeral fora.
  contact:
    name: WuzzChat Core Team
    url: https://chat.wuzzhub.id
    email: support@wuzzhub.id
  license:
    name: Proprietary / MIT
servers:
  - url: https://chat-api.wuzzhub.id
    description: Live Production Server (VPS)
  - url: http://localhost:8080
    description: Local Development Server
...
paths:
...
  /api/v1/auth/provision-token:
    post:
      tags:
        - B2B Gateway
      summary: Provision JIT User & Issue 60s Exchange Token
      description: |
        Called server-to-server by third-party B2B backend systems to silently provision or upsert a user within their tenant and obtain a single-use 60-second exchange token.
        
        Requires `X-App-ID` and `X-App-Secret` headers matching an active tenant.
      security:
        - B2BAppID: []
          B2BAppSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisionTokenRequest'
      responses:
        '200':
          description: Exchange token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionTokenResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Missing or invalid B2B application credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Tenant is inactive, suspended, or rate-limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/exchange:
    post:
      tags:
        - B2B Gateway
      summary: Exchange One-Time Token for Session JWT
      description: |
        Called by the third-party client application (Web/Mobile SDK) using the short-lived exchange token to receive a 30-day WuzzChat session JWT.
      parameters:
        - $ref: '#/components/parameters/HeaderTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeTokenRequest'
      responses:
        '200':
          description: Token exchanged successfully. Returns standard session JWT and user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Invalid request or missing exchange token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Exchange token expired, consumed, or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # =========================================================================
  # 2. AUTHENTICATION & IDENTITY
  # =========================================================================
  /api/auth/register:
    post:
      tags:
        - Authentication & Identity
      summary: Register a New User Account
      description: |
        Registers a new user account within the resolved tenant. Rate-limited to 15 requests/minute per IP.
      parameters:
        - $ref: '#/components/parameters/HeaderTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Validation error or invalid inputs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Username already exists in this tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/login:
    post:
      tags:
        - Authentication & Identity
      summary: Authenticate User & Issue Session JWT
      description: |
        Validates username and password within the tenant. Enforces a maximum of 2 active devices per user.
      parameters:
        - $ref: '#/components/parameters/HeaderTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Too many failed login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/me:
    get:
      tags:
        - Authentication & Identity
      summary: Get Current User Profile
      description: Returns the authenticated user profile along with tenant information.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/refresh:
    post:
      tags:
        - Authentication & Identity
      summary: Refresh Token (Sliding Renewal)
      description: >
        Issues a new JWT (new JTI, full 30-day lifetime) when the current token has less than 50% of its lifetime
        left. Otherwise returns `refreshed=false` and the client keeps its token. Sessions are capped at 365 days
        since the original login (`auth_time`). The old token is left to expire naturally; its session record is
        revoked. Revoked/expired tokens are rejected with 401 before reaching this handler.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                type: object
                properties:
                  refreshed:
                    type: boolean
                  token:
                    type: string
                    description: Present only when refreshed=true
                  expires_at:
                    type: string
                    format: date-time
        '401':
          description: Token invalid/revoked, session inactive, or absolute session age (365 days) exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/logout:
    post:
      tags:
        - Authentication & Identity
      summary: Logout & Revoke Current Session
      description: Revokes the current JWT session and blacklists its JTI token.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/sessions:
    get:
      tags:
        - Authentication & Identity
      summary: List Active Sessions
      description: Retrieves all active sessions for the authenticated user across devices.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/auth/sessions/revoke-others:
    post:
      tags:
        - Authentication & Identity
      summary: Revoke All Other Active Sessions
      description: Terminates all active sessions except the current caller session.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Other sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  revoked_count:
                    type: integer
                    example: 1
  /api/auth/devices:
    get:
      tags:
        - Authentication & Identity
      summary: List Registered Devices
      description: Returns list of devices registered under the user account.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Devices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
...
  /api/auth/profile:
    put:
      tags:
        - Authentication & Identity
      summary: Update User Profile
      description: Updates display name, status bio, and avatar image URL.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  example: Alice Smith
                status:
                  type: string
                  example: Available for meetings
                avatar_url:
                  type: string
                  example: /uploads/avatar-alice.png
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/auth/verify-password:
    post:
      tags:
        - Authentication & Identity
      summary: Verify User Password
      description: Pre-flight check to verify current password before high-privilege operations.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
                  example: secret12345
      responses:
        '200':
          description: Password verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
        '400':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/change-password:
    post:
      tags:
        - Authentication & Identity
      summary: Change Password
      description: Updates user password and revokes all active sessions except the current one.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Standard 30-day session token passed as Authorization: Bearer <token>."
    B2BAppID:
      type: apiKey
      in: header
      name: X-App-ID
      description: Third-party tenant application ID.
    B2BAppSecret:
      type: apiKey
      in: header
      name: X-App-Secret
      description: Third-party tenant secret key.
...